// admin/src/api/generated/schemas/adminAdminSwapTransactionItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminAdminSwapTransactionItem = {
  id?: string;
  project_name?: string;
  wallet_address?: string;
  chain_id?: string;
  chain_name?: string;
  provider?: string;
  from_token_symbol?: string;
  from_token_address?: string;
  from_token_decimals?: number;
  to_token_symbol?: string;
  to_token_address?: string;
  to_token_decimals?: number;
  from_amount?: string;
  to_amount?: string;
  tx_hash?: string;
  status?: string;
  fee_usd?: string;
  gas?: string;
  created_at?: string;
  broadcasted_at?: string;
};